+++ ft_strlen.c
@@ -6,11 +6,13 @@
 /*   By: yowoo <yowoo@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/09 11:20:56 by yowoo             #+#    #+#             */
-/*   Updated: 2023/10/17 18:47:28 by yowoo            ###   ########.fr       */
+/*   Updated: 2023/10/26 15:02:08 by yowoo            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-unsigned long	ft_strlen(const char *str)
+#include "libft.h"
+
+size_t	ft_strlen(const char *str)
 {
 	unsigned int	len;
